Update Elasticsearch in a background task.
