+++ aspnetcore/src/api/Background/BackgroundElasticsearchPersonUpdateQueue.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Concurrent;
+using System.Threading;
+using System.Threading.Tasks;
+
+public class BackgroundElasticsearchPersonUpdateQueue
+{
+    private ConcurrentQueue<Func<CancellationToken, Task>> _workItems = new ConcurrentQueue<Func<CancellationToken, Task>>();
+    private SemaphoreSlim _signal = new SemaphoreSlim(0);
+
+    public async Task<Func<CancellationToken, Task>> DequeueAsync(CancellationToken cancellationToken)
+    {
+        await _signal.WaitAsync(cancellationToken);
+        _workItems.TryDequeue(out var workItem);
+
+        return workItem;
+    }
+
+    public void QueueBackgroundWorkItem(Func<CancellationToken, Task> workItem)
+    {
+        if (workItem == null)
+        {
+            throw new ArgumentNullException(nameof(workItem));
+        }
+
+        _workItems.Enqueue(workItem);
+        _signal.Release();
+    }
+}
